+++ packages/gamut-templates/src/index.ts
@@ -1,2 +1,3 @@
 export { default as Interstitial } from './Interstitial';
+export { default as GridForm } from './GridForm';
 export { default as SplitInterstitial } from './SplitInterstitial';